+++ CHANGELOG.md
@@ -11,6 +11,16 @@ Features:
 * Configuration file contains the walltime taken by each run
 * It is now possible to upload or download any file via its full path
 
+1.0.8 (???)
+-----------
+
+Behavior changes:
+* No longer default to overwriting trace directories. ReproZip will ask what to do or exit with an error if one of --continue/--overwrite is not provided
+
+Bugfixes:
+* Fix an issue identifying Debian packages when a file's in two packages
+* Fix Python error `Mixing iteration and read methods would lose data`
+
 1.0.7 (2016-08-22)
 ------------------